+++ app/serializers/pathfinder2/character_serializer.rb
@@ -9,10 +9,10 @@ module Pathfinder2
                :avatar, :weapon_skills, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions,
                :ability_boosts_v2, :skill_boosts, :attacks, :provider, :lores, :selected_features, :spells_info, :class_dc,
                :spell_attack, :spell_dc, :spent_spell_slots, :formatted_static_spells, :modified_abilities, :spell_list,
-               :can_have_pet, :can_have_familiar, :raw_abilities, :experience, :max_dying
+               :can_have_pet, :can_have_familiar, :raw_abilities, :experience, :max_dying, :speeds
 
     delegate :health, :abilities, :skills, :subclasses, :saving_throws_value, :weapon_skills, :raw_abilities, :max_dying,
-             :money, :lores, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions, :attacks,
+             :money, :lores, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions, :attacks, :speeds,
              :features, :spells_info, :class_dc, :spell_attack, :spell_dc, :formatted_static_spells, :modified_abilities,
              :can_have_pet, :can_have_familiar, to: :decorator
     delegate :name, :id, :data, :created_at, to: :object